pyproject.toml
[project]
name = "celery-playground"
version = "0.1.0"
description = "Celery Production Mastery - Complete Engineering Handbook"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "celery-types>=0.24.0",
    "celery-once>=3.0.1",
    "django>=6.0.2",
    "django-celery-beat>=2.7.0",
    "django-celery-results>=2.5.1",
    "redis>=7.1.1",
    "flower>=2.0.1",
    "prometheus-client>=0.20.0",
    "python-json-logger>=2.0.7",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-celery>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "requests>=2.32.0",
    "pillow>=10.0.0",
    "psutil>=7.2.2",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-django>=4.8.0",
    "pytest-celery>=1.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "ipython>=8.26.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["celery_playground*", "demo*", "production_patterns*"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "celery_playground.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py313"
